compliance_report_generator/pyproject.toml
[project]
name = "compliance-report-generator"
version = "0.1.0"
description = ""
authors = [
    {name = "vapmail16",email = "vapmail16@gmail.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "langchain (>=0.3.21,<0.4.0)",
    "openai (>=1.69.0,<2.0.0)",
    "qdrant-client (>=1.13.3,<2.0.0)",
    "neo4j (>=5.28.1,<6.0.0)",
    "streamlit (>=1.44.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pdfplumber (>=0.11.6,<0.12.0)",
    "langchain-community (>=0.3.20,<0.4.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "langchain-openai (>=0.3.11,<0.4.0)",
    "pydantic (>=2.11.1,<3.0.0)",
    "langgraph (>=0.3.21,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"


[tool.poetry]
package-mode = false